I think the internals are pretty much there. Just the CLI to put in.

+++ class/Server.js
@@ -1,6 +1,7 @@
 "use strict";
 
 var core = require('jscore');
+var net = require('net');
 var constants = require('./Constants.js');
 var SNI = require('./SNI.js');
 var HTTP = require('./HTTP.js');
@@ -16,6 +17,9 @@ var Server = module.exports = core.Class.extend(function()
 })
 .implement({
 	_clientIndex: 0,
+	_404: true,
+	_500: true,
+	_504: true,
 	listen: core.fn.overload(
 	{
 		// TCP/IP
@@ -57,14 +61,41 @@ var Server = module.exports = core.Class.extend(function()
 			["string", "regex"],
 			{ type: "object" }
 		],
-		call: function(hostname, options)
+		call: function(hostname, upstream)
 		{
 			this._routes.push({
 				hostname: hostname,
 				rx: typeof hostname === 'string' ? this._globToRegex(hostname) : hostname,
-				options: core.util.combine({}, options, { keepHalfOpen: true })
+				upstream: core.util.combine({}, upstream, { allowHalfOpen: true })
 			});
 
+			return this;
+		}
+	}),
+	set404: core.fn.overload(
+	{
+		args: [['boolean', "object", "function"]],
+		call: function(value)
+		{
+			this._404 = value;
+			return this;
+		}
+	}),
+	set500: core.fn.overload(
+	{
+		args: [['boolean', "object", "function"]],
+		call: function(value)
+		{
+			this._500 = value;
+			return this;
+		}
+	}),
+	set504: core.fn.overload(
+	{
+		args: [['boolean', "object", "function"]],
+		call: function(value)
+		{
+			this._504 = value;
 			return this;
 		}
 	}),
@@ -99,7 +130,7 @@ var Server = module.exports = core.Class.extend(function()
 	},
 	_listen: function(secure, args)
 	{
-		var server = core.net.createServer({ allowHalfOpen: true });
+		var server = net.createServer({ allowHalfOpen: true });
 
 		server
 			.on('listening', this._onListening.bind(this, server))
@@ -124,7 +155,7 @@ var Server = module.exports = core.Class.extend(function()
 	},
 	_onListening: function(server)
 	{
-		this.emit('listening', server);
+		core.fn.safe( this, 'emit', 'listening', server);
 	},
 	_onConnection: function(server, client)
 	{
@@ -135,11 +166,11 @@ var Server = module.exports = core.Class.extend(function()
 
 		client.once('data', this._onData.bind(this, server, client));
 
-		this.emit('connection', client, server);
+		core.fn.safe( this, 'emit', 'connection', client, server);
 	},
 	_onData: function(server, client, firstPacket)
 	{
-		var hostname = false;
+		var hostname = '';
 
 		try
 		{
@@ -150,28 +181,40 @@ var Server = module.exports = core.Class.extend(function()
 		}
 		catch (err)
 		{
-			if (!(err instanceof SNI.NotPresent))
+			if (!(err instanceof SNI.NotPresent) && !(err instanceof HTTP.MissingHostHeader))
 			{
-				client.emit('warning', err.message ? err.message : err);
+				core.fn.safe(client, 'emit', 'warning', err.message ? err.message : err);
 				client.destroy();
 
 				return;
 			}
 		}
 
-		this._proxy(server, client, firstPacket, hostname);
+		var upstream = this._resolveRoute(hostname);
+
+		if (!upstream)
+		{
+			if (client.secure)
+				client.destroy();
+			else
+				this._on404(server, client, firstPacket);
+		}
+		else
+		{
+			this._proxy(server, client, firstPacket, upstream);
+		}
 	},
 	_onProxy: function(proxy)
 	{
-		this.emit('proxy', proxy);
+		core.fn.safe( this, 'emit', 'proxy', proxy);
 	},
 	_onClose: function(server)
 	{
-		this.emit('close', server);
+		core.fn.safe( this, 'emit', 'close', server);
 	},
 	_onError: function(server, err)
 	{
-		this.emit('error', err, server);
+		core.fn.safe( this, 'emit', 'error', err, server);
 	},
 	_speculativeTimeout: function(client)
 	{
@@ -181,9 +224,8 @@ var Server = module.exports = core.Class.extend(function()
 
 		var timeout = setTimeout(function()
 		{
-			client
-				.emit('warning', 'no data received')
-				.destroy();
+			core.fn.safe(client, 'emit', 'warning', 'no data received');
+			client.destroy();
 		}, 1000);
 
 		client.once('data', function()
@@ -191,31 +233,46 @@ var Server = module.exports = core.Class.extend(function()
 			clearTimeout(timeout);
 		});
 	},
-	_proxy: function(server, client, firstPacket, hostname)
+	_proxy: function(server, client, firstPacket, upstream, silent)
 	{
-		var route = this._resolveRoute(hostname);
-
-		if (!route)
-		{
-			if (client.secure)
-				client.destroy();
-			else
-				client.end(constants.responseNoRoute, 'utf8');
-
-			return;
-		}
-
 		var proxy = Proxy.create(server, client);
 
-		proxy.upstream.on('error', function(err)
-		{
-			if (client.bytesWritten === 0 && !client.secure)
-				proxy.client.end(constants.responseUpstreamError, 'utf8');
-		});
+		this._proxies.push(proxy);
+
+		proxy
+			.on('close', function()
+			{
+				var i = this._proxies.length;
+				while (i--)
+				{
+					if (this._proxies[i] === proxy)
+					{
+						this._proxies.splice(i, 1);
+						break;
+					}
+				}
+			}.bind(this))
+			.on('error', function()
+			{
+				if (!client.secure && !silent)
+					this._on504(server, client, firstPacket);
+			}.bind(this));
 
 		this._onProxy(proxy);
 
-		proxy.connect(route, firstPacket);
+		try
+		{
+			proxy.connect(upstream, firstPacket);
+		}
+		catch (err)
+		{
+			if (client.secure)
+				client.destroy();
+			else if (!silent)
+				this._on500(server, client, firstPacket);
+
+			this._onError(null, err);
+		}
 	},
 	_resolveRoute: function(hostname)
 	{
@@ -224,10 +281,43 @@ var Server = module.exports = core.Class.extend(function()
 		while (i--)
 		{
 			if (this._routes[i].rx.test(hostname))
-				return this._routes[i].options;
+				return this._routes[i].upstream;
 		}
 
 		return false;
+	},
+	_on404: function(server, client, firstPacket)
+	{
+		if (this._404 === false)
+			return;
+		else if (this._404 === true)
+			client.end(constants.responseNoRoute, 'utf8');
+		else if (this._404 instanceof Function)
+			this._404(server, client, firstPacket);
+		else
+			this._proxy(server, client, firstPacket, this._404, true);
+	},
+	_on500: function(server, client, firstPacket)
+	{
+		if (this._500 === false)
+			return;
+		else if (this._500 === true)
+			client.end(constants.responseUpstreamInvalid, 'utf8');
+		else if (this._500 instanceof Function)
+			this._500(server, client, firstPacket);
+		else
+			this._proxy(server, client, firstPacket, this._500, true);
+	},
+	_on504: function(server, client, firstPacket)
+	{
+		if (this._504 === false)
+			return;
+		else if (this._504 === true)
+			client.end(constants.responseUpstreamError, 'utf8');
+		else if (this._504 instanceof Function)
+			this._504(server, client, firstPacket);
+		else
+			this._proxy(server, client, firstPacket, this._504, true);
 	}
 });
